src/board/uos3-proto/: Add GPIO driver, use in LED, WDT drivers.

+++ src/board/uos3-proto/tm4c_startup_gcc.c
@@ -278,7 +278,7 @@ ResetISR(void)
     // this project.
     //
     HWREG(NVIC_CPAC) = ((HWREG(NVIC_CPAC) &
-                         ~(NVIC_CPAC_CP10_M | NVIC_CPAC_CP11_M)) |
+                        (uint32_t)~(NVIC_CPAC_CP10_M | NVIC_CPAC_CP11_M)) |
                         NVIC_CPAC_CP10_FULL | NVIC_CPAC_CP11_FULL);
 
     //
